+++ settlement/local/local.go
@@ -1,4 +1,4 @@
-package mock
+package local
 
 import (
 	"context"
